==> notebooks/1.3.Uniformar_satelitalesCSV(NASA).ipynb <==
# Uniformar los datos satelitales de NASA - diarios por estación

import pandas as pd
import numpy as np
import os

### Funciones

'''Definir periodo de los datos'''
fecha_ini = '1981-01-01'
fecha_fin = '2020-12-31'

'''devolver la fecha en base al año y el numero de día'''
def aniodia(anio, dia=1):
    fecha = pd.to_datetime(dia-1, unit='D', origin=str(anio))
    return fecha.date()

'''devolver el rango de fecha en base al año, dia inicial y periodo'''
def rango_fecha1(anio, dia, periodo):
    rango = pd.date_range(start = aniodia(anio, dia), freq = 'D', periods = periodo)
    return rango

'''leer los archivos en un carpeta dada SI son archivos CSV y si su nombre comienza con NASA'''
def leer_dir(path):
    with os.scandir(path) as ficheros:
        ficheros = [fichero.name for fichero in ficheros if fichero.is_file() \
                    and fichero.name.endswith('.csv') and fichero.name.startswith('nasa')]
    ficheros.sort()
    return ficheros

'''crear un DF con una sola columna en base a otro DF y una columna del mismo'''
def creardf(data, col):
    dat = pd.DataFrame()
    dat[col] = data[col]
    return dat

'''guardar un DF en base a un nombre y un directorio'''
def guardar(data, nom, path='../datos/1.0.variables/'):
    pth = path + nom
    data.to_csv(pth, index = False)

### Generar archivos con datos uniformados 

dir_f4 = '../datos/1.3.formato5_nasa/'
archivos = leer_dir(dir_f4)


for p in archivos:
    ruta = dir_f4 + p
    data = pd.read_csv(ruta, skiprows=13)
    
    '''leer anio y dia de inicio'''
    anio = data.iloc[0,0]
    dia = data.iloc[0,1]
    dias = data.shape[0]
    
    '''adicionar la columna fecha'''
    data['fecha'] = rango_fecha1(anio, dia, dias)
    
    '''cambiar de nombre las columnas'''
    data = data.rename(columns={"fecha": "fecha", "PRECTOTCORR": "pp", "T2M_MAX":\
                                "tmax", "T2M_MIN": "tmin", "T2M": "tmed", "RH2M": "hmed"})

    '''reordenar las columnas y eliminar columnas YEAR y DOY'''
    data = data.reindex(['fecha','pp','tmax','tmin','tmed','hmed'], axis=1)    
    
    '''periodo de datos'''
    data = data[(data['fecha'] >= fecha_ini) & (data['fecha'] <= fecha_fin)]
    
    '''guardar archivo'''
    dir_guardar = '../datos/1.0.formato_csv_corregido/' + p
    data.to_csv(dir_guardar, index = False)
    
    

### Generar archivos para cada variable meteorológica 

'''llamar a la función LEER_DIR() para recuperar archivos de un directorio'''
lista_arch = leer_dir('../datos/1.0.formato_csv_corregido/')

'''leer un DF en forma temporal'''
pin = '../datos/1.0.formato_csv_corregido/' + lista_arch[0]
temp = pd.read_csv(pin)

'''crear DF para cada variable, tomando como base el DF TEMP'''
dfpp = creardf(temp,['fecha'])
dftmx = creardf(temp,['fecha'])
dftmn = creardf(temp,['fecha'])
dftmd = creardf(temp,['fecha'])
dfhmd = creardf(temp,['fecha'])

'''crear una lista con los nombres de las columnas'''
nom_col = ['fecha']

for p in lista_arch:
    pin = '../datos/1.0.formato_csv_corregido/' + p
    file = pd.read_csv(pin)
    dfpp = pd.concat([dfpp, file['pp']], axis=1)
    dftmx = pd.concat([dftmx, file['tmax']], axis=1)
    dftmn = pd.concat([dftmn, file['tmin']], axis=1)
    dftmd = pd.concat([dftmd, file['tmed']], axis=1)
    dfhmd = pd.concat([dfhmd, file['hmed']], axis=1)
    
    cad = p.split('_')
    nom_col.append(str(cad[1]))
    
dfpp.columns = nom_col
dftmx.columns = nom_col
dftmn.columns = nom_col
dftmd.columns = nom_col
dfhmd.columns = nom_col

guardar(dfpp, 'nasa_diario_pp.csv')
guardar(dftmx, 'nasa_diario_tmax.csv')
guardar(dftmn, 'nasa_diario_tmin.csv')
guardar(dftmd, 'nasa_diario_tmed.csv')
guardar(dfhmd, 'nasa_diario_hmed.csv')

'''Dataframe precipitación con datos de NASA'''
print(dfpp.head(5))
