==> notebooks/1.4.Uniformar_satelitalesNOAA   .ipynb <==
# Uniformar Datos Satelitales NOAA

import pandas as pd
import numpy as np

'''Cargar datos de NOAA '''
oni=pd.read_csv('../datos/1.4.formato6_noaa/noaa_oni.csv', sep = ',', parse_dates=True)
oni.head()

'''Aplicar Melt '''
oni1=pd.melt(oni,id_vars=["Year"],value_vars=["12","1","2","3","4","5","6","7","8","9","10","11"],value_name="oni").sort_values(by="Year").reset_index(drop=True)
oni1['fecha']=pd.to_datetime(oni1['Year'].astype(str)+'-'+oni1['variable'].astype(str))
oni1.head()

oni1.drop(['Year','variable'],axis=1,inplace=True)
oni1=oni1.sort_values(by="fecha").reset_index(drop=True)
oni1= oni1[['fecha','oni']]
oni1.to_csv('../datos/1.0.variables/noaa_mensual_oni.csv', index = False)

print(oni1.head(10))
