==> notebooks/2.0 Generar datos para Climatol.ipynb <==
# Generar datos para Climatol

'''instalar openpyxl para abrir varias hojas de un libro de excel'''
# %pip install openpyxl

import pandas as pd
import numpy as np
import datetime
from calendar import monthrange

## Generar archivos txt para cada estacion

path_f03 = '../datos/1.0.Estaciones/estaciones_senamhi_generado.csv'
df03 = pd.read_csv(path_f03)

for p in df03['archivo_csv_corregido']:
    file = pd.read_csv('../datos/1.0.formato_csv_corregido/'+p)
    file['fecha']=pd.to_datetime(file['fecha'])
    file['anio']=file['fecha'].dt.year
    file['mes']=file['fecha'].dt.month
    file['dia']=file['fecha'].dt.day
    file=file.reindex(columns = ['anio','mes','dia','pp','tmax','tmin','tmed','hmed'])
    path_a = p.split('_')
    path_a2 = path_a[1]
    path_f = '../datos/2.1.estaciones_climatol/' + path_a2 + '.txt'    
    file.to_csv(path_f, na_rep='NA', index=False, header=False)

## Generar el archivo stations.txt

path_f = '../datos/1.0.Estaciones/estaciones_senamhi_generado.csv'
df_sta = pd.read_csv(path_f)

cols_0 = ['archivo', 'lon', 'lat', 'altitud', 'cod', 'nombre']
df_s = pd.DataFrame(columns=cols_0)
df_s['lon'] = df_sta['lon']
df_s['lat'] = df_sta['lat']
df_s['altitud'] = df_sta['altitud']
df_s['cod'] = df_sta['id']

t = df_sta.shape[0]
for i in range (0, t):
    df_s.loc[i,'archivo'] = chr(34) + str(df_sta.loc[i,'id']) + '.txt' + chr(34)
    df_s.loc[i,'nombre'] = chr(34) + df_sta.loc[i,'estacion'] + chr(34)

df_s.to_csv('../datos/2.1.estaciones_climatol/stations.txt', index=False, header=False)

'''Reemplazar comillas en el txt'''
