==> notebooks/2.1 Homogenizar Climatol R.ipynb <==
# HOMOGENIZAR DATOS CON CLIMATOL - R

#Instalar Climatol -  solo instalar una vez
#install.packages("climatol")

#Cargar climatol
library(climatol)

#Configurar ruta del archivo
setwd("../datos/2.1.estaciones_climatol")
getwd()

anio_min<-1981 #año mínimo para el análisis
anio_max<-2020 #año máximo para el análisis

##  PARA PRECIPITACION

#1.Convertir los datos a formato climatol 
daily2climatol(stfile = "stations.txt", stcol = 1:6, datcol = c(1:3,4), varcli="pre",anyi = anio_min,anyf = anio_max, mindat = 1095, sep = ",", na.strings = "NA")

#2.Aplicar homogen en modo exploratorio diario
homogen("pre",anio_min, anio_max, expl = TRUE)

#3.Transformar los datos diarios a datos mensuales  valm = 1 para precipitaci?n
dd2m(varcli = "pre",anio_min,anio_max,valm = 1)

#4.Aplicar homogen en modo exploratorio mensual
homogen("pre-m",anio_min, anio_max, expl = TRUE)

#5.Ajuste mensual  std = 2 (El valor 2 se utiliza para variables como precipitaci?n)
#dz.min y dz.max Histogram of normalized anomalies 
#Histogram of maximum windowed SNHT para snht1  ,  Histogram of maximum global SNHT  snht2
homogen("pre-m",anio_min, anio_max, dz.min = -9, dz.max = 10, snht1 = 30, snht2 = 44, std =2, cutlev = 2.3, vmin = 0 )

#6.Ajuste diario (metad = TRUE toma los break mensuales)
homogen("pre",anio_min, anio_max,  dz.min = -16, dz.max = 26, snht1 = 220, snht2 = 250, std =2, vmin = 0, metad = TRUE)

#7.Generar las series homogenizadas diarias
dahstat('pre',anio_min,anio_max, stat = 'series')

#8.Generar las series homogenizadas mensuales
dahstat('pre-m',anio_min,anio_max, stat = 'series')

##  PARA TEMPERATURA MAXIMA

#1.Convertir los datos a formato climatol 
daily2climatol(stfile = "stations.txt", stcol = 1:6, datcol = c(1:3,5), varcli="tmax",anyi = anio_min ,anyf = anio_max, mindat = 365, sep = ",", na.strings = "NA")

#2.Aplicar homogen en modo exploratorio diario
homogen("tmax",anio_min, anio_max, expl = TRUE)

#3.Transformar los datos diarios a datos mensuales  valm = 2 para temperatua maxima
dd2m(varcli = "tmax", anio_min,anio_max,valm = 2)

#4.Aplicar homogen en modo exploratorio mensual
homogen("tmax-m",anio_min, anio_max, expl = TRUE)

#5.Ajuste mensual  std =3 (El valor 3 se utiliza para variables como temperatura m?xima y m?nima)
homogen("tmax-m",anio_min, anio_max, dz.min = -8, dz.max = 6, snht1 = 80, snht2 = 170, std = 3, vmin = -10, cutlev = 2.3)

#6.Ajuste diario (metad = TRUE toma los break mensuales)
homogen("tmax",anio_min, anio_max, dz.min = -30, dz.max = 50, snht1 = 500, snht2 = 3500,std =3, vmin = -10, vmax = 30, metad = TRUE)

#7.Generar las series homogenizadas diarias
dahstat('tmax',anio_min,anio_max, stat = 'series')

#8.Generar las series homogenizadas mensuales
dahstat('tmax-m',anio_min,anio_max, stat = 'series')

## PARA TEMPERATURA MINIMA

#1.Convertir los datos a formato climatol 
daily2climatol(stfile = "stations.txt", stcol = 1:6, datcol = c(1:3,6), varcli="tmin",anyi = anio_min ,anyf = anio_max, mindat = 365, sep = ",", na.strings = "NA")

#2.Aplicar homogen en modo exploratorio diario
homogen("tmin",anio_min, anio_max, expl = TRUE)

#3.Transformar los datos diarios a datos mensuales  valm = 2 (mean) para temperatura
dd2m(varcli = "tmin", anio_min,anio_max,valm = 2)

#4.Aplicar homogen en modo exploratorio mensual
homogen("tmin-m",anio_min, anio_max, expl = TRUE)

#5.Ajuste mensual  std =3 (El valor 3 se utiliza para variables como temperatura m?xima y m?nima)
homogen("tmin-m",anio_min, anio_max, dz.min = -5, dz.max = 6, snht1 = 80, snht2 = 170, std = 3, vmin = -20 , cutlev = 2.3)

#6.Ajuste diario (metad = TRUE toma los break mensuales)
homogen("tmin",anio_min, anio_max, dz.min = -13, dz.max = 12, snht1 = 400, snht2 = 2400, std =3, vmin = -20, vmax = 30 , metad = TRUE)

#7.Generar las series homogenizadas diarias
dahstat('tmin',anio_min,anio_max, stat = 'series')

#8.Generar las series homogenizadas mensuales
dahstat('tmin-m',anio_min,anio_max, stat = 'series')

## PARA TEMPERATURA MEDIA

anio_min<-1991 #año mínimo para el análisis
anio_max<-2020 #año máximo para el análisis

#1.Convertir los datos a formato climatol 
daily2climatol(stfile = "stations.txt", stcol = 1:6, datcol = c(1:3,7), varcli="tmed",anyi = anio_min ,anyf = anio_max, mindat = 365, sep = ",", na.strings = "NA")

#2.Aplicar homogen en modo exploratorio diario
homogen("tmed",anio_min, anio_max, expl = TRUE)

#3.Transformar los datos diarios a datos mensuales  valm = 2 (mean) para temperatura
dd2m(varcli = "tmed", anio_min,anio_max,valm = 2)

#4.Aplicar homogen en modo exploratorio mensual
homogen("tmed-m",anio_min, anio_max, expl = TRUE)

#5.Ajuste mensual  std =3 (El valor 3 se utiliza para variables como temperatura m?xima y m?nima)
homogen("tmed-m",anio_min, anio_max, dz.min = -5, dz.max = 6, snht1 = 80, snht2 = 170, std = 3, vmin = -20 , cutlev = 2.3)

#6.Ajuste diario (metad = TRUE toma los break mensuales)
homogen("tmed",anio_min, anio_max, dz.min = -10, dz.max = 12, snht1 = 400, snht2 = 2200, std =3, vmin = -20, vmax = 30 , metad = TRUE)

#7.Generar las series homogenizadas diarias
dahstat('tmed',anio_min,anio_max, stat = 'series')

#8.Generar las series homogenizadas mensuales
dahstat('tmed-m',anio_min,anio_max, stat = 'series')
