==> notebooks/3.2. Resumen_anual.ipynb <==
# Resumen anual

import os
import pandas as pd


fecha_ini = '1981-01-01'
fecha_fin = '2020-12-31'


'''leer los archivos en un carpeta dada SI son archivos CSV'''
def leer_dir(path, cond1 = 'indice'):
    with os.scandir(path) as ficheros:
        ficheros = [fichero.name for fichero in ficheros if fichero.is_file() 
                                                            and fichero.name.endswith('12.csv') 
                                                            and fichero.name.count(cond1)>0]
    ficheros.sort()
    return ficheros

def procesar(nom):
    
    if (nom.count('indices') > 0):
        ruta = '../datos/3.0.spi_spei/'
    else:
        ruta = '../datos/1.0.variables/'   
    data = pd.read_csv(ruta + nom)
    
    data['fecha'] = pd.to_datetime(data['fecha'])
    data = data[(data['fecha'] >= fecha_ini) & (data['fecha'] <= fecha_fin)]
    
    '''Para variable precipitación se obtiene la media de la suma de todos los meses'''
    if (nom.count('pp') > 0): 
        '''Para precipitación'''
        return data.groupby(pd.Grouper(key='fecha', freq = 'YS')).sum()
    elif (nom.count('indice') > 0) :
        '''Para spi-spei'''
        return data.groupby(pd.Grouper(key='fecha', freq = 'YS')).mean()
    else:
        '''Para temperatura y humedad'''
        return data.groupby(pd.Grouper(key='fecha', freq = 'YS')).mean()


def adicionar_serie(s, df, col):
    df.sort_values(by='id', inplace=True, ascending=True)
    s.sort_index(ascending=True, inplace=True)
    c = nombre_col(col)
    df[c] = s.values
    return df
    
def nombre_col(cadena):
    cadena = cadena.split('.')
    cadena = cadena[0].split('_')
    cad = cadena[2] + '_media_anual'
    return cad

dir_est = '../datos/1.0.Estaciones/estaciones_senamhi_generado.csv'
cols = ['id', 'lon', 'lat', 'altitud', 'estacion']

estaciones = pd.read_csv(dir_est, usecols=cols)[cols]


## Generar resumen media anual en base a los datos CHIRPS, NASA y OBSERVADOS

fuente_archivos = ['chirps_mensual_pp_bc.csv','nasa_mensual_hmed.csv',
                   'nasa_mensual_tmax_bc.csv','nasa_mensual_tmed.csv','nasa_mensual_tmin_bc.csv',
                   'indices_spei1.csv','indices_spi1.csv',
                   'indices_spei3.csv','indices_spi3.csv',
                   'indices_spei6.csv','indices_spi6.csv',
                   'indices_spei12.csv','indices_spi12.csv']

for archivo in fuente_archivos:
    serie = procesar(archivo)     
    serie['fecha']=pd.DatetimeIndex(serie.index)
    serie['anio']=pd.DatetimeIndex(serie['fecha']).year
    serie.drop('fecha', axis=1, inplace=True)
    
    if (archivo.count('chirps_mensual'))>0:
        archivo = archivo.replace('chirps_mensual_','')
    elif (archivo.count('nasa_mensual'))>0:
        archivo = archivo.replace('nasa_mensual_','')
    else:
        archivo = archivo.replace('indices_','')
    serie.to_csv('../datos/1.0.variables/resumen_anual_'+archivo, index = False)  


serie
